simplify data fetching and UI
