+++ BiggerNotes/UI/NoteList.swift
@@ -12,20 +12,33 @@ struct NoteList: View {
     @EnvironmentObject var appSettingsViewModel: AppSettingsViewModel
     @EnvironmentObject var noteViewModel: NoteViewModel
     @EnvironmentObject var router: Router
-    @State private var refreshId = UUID()
-
+    
+    @SectionedFetchRequest<Bool, Note>(
+        sectionIdentifier: \.favorite,
+        sortDescriptors: [
+            NSSortDescriptor(keyPath: \Note.favorite, ascending: false),
+            NSSortDescriptor(keyPath: \Note.modified, ascending: false)
+        ]
+    )
+    var sectionedNotes: SectionedFetchResults<Bool, Note>
+    
     var body: some View {
         NavigationStack(path: $router.path) {
             List {
-                NoteListSection(notes: noteViewModel.favoriteNotes, sectionHeaderText: "Favorites", sectionHeaderIcon: "star.fill", expanded: $appSettingsViewModel.expandFavoritesSection)
-                NoteListSection(notes: noteViewModel.nonfavoriteNotes, sectionHeaderText: "Notes", sectionHeaderIcon: "note.text", expanded: $appSettingsViewModel.expandNotesSection)
+                ForEach(sectionedNotes) { section in
+                    Section {
+                        ForEach(section) { note in
+                            NoteListItem(note: note)
+                        }
+                    } header: {
+                        Text(section.id ? "Favorites" : "Notes")
+                    }
+                }
             }
-            .task {
-                noteViewModel.prune()
-            }
+            .listStyle(SidebarListStyle())
             .navigationBarTitle("Notes")
             .navigationDestination(for: Note.self) { note in
-                NoteDetail(note: note, parentRefreshId: $refreshId)
+                NoteDetail(note: note)
             }
             .toolbar {
                 ToolbarItemGroup(placement: .navigationBarLeading) {
@@ -41,7 +54,6 @@ struct NoteList: View {
             }
             .errorAlert(errorMessage: $noteViewModel.errorMessage)
         }
-        .id(refreshId)
     }
 }
